Renamed slices directory and rate slice file
