+++ src/reducers.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 
-import rateReducer from './slices/rateSlice';
+import rateReducer from './Slices/rate';
 
 export default combineReducers({
   rate: rateReducer
